Rename generators to match planned module names
