+++ src/generators/float.js
