+++ src/generators/fontWeights.js
