+++ src/generators/minHeight.js
@@ -1,0 +1,14 @@
+import _ from 'lodash'
+import defineClass from '../util/defineClass'
+
+function defineMinHeights(heights) {
+  return _.map(heights, (size, modifer) => {
+    return defineClass(`min-h-${modifer}`, {
+      'min-height': `${size}`,
+    })
+  })
+}
+
+export default function(config) {
+  return _.flatten([defineMinHeights(config.minHeight)])
+}
